+++ src/view/response.py
@@ -1,18 +1,25 @@
+"""
+view.py public response APIs
+
+This module contains the `Response` class, which is conventionally used as the base response in view.py
+All other classes that inherit from it are contained in this module.
+"""
 from __future__ import annotations
 
+import uuid
+from contextlib import suppress
 from datetime import datetime as DateTime
 from pathlib import Path
 from typing import Any, Dict, Generic, TextIO, TypeVar, Union
 
 import aiofiles
 import ujson
+from typing_extensions import final
 
 from .components import DOMNode
 from .exceptions import InvalidResultError
 from .typing import BodyTranslateStrategy, ResponseBody, SameSite, ViewResult
 from .util import timestamp
-from contextlib import suppress
-import uuid
 
 T = TypeVar("T")
 
@@ -23,7 +30,14 @@ HTMLContent = Union[TextIO, str, Path, DOMNode]
 
 
 class Response(Generic[T]):
-    """Wrapper for responses."""
+    """
+    Base view.py response class.
+    
+    Technically speaking, it's not required for an object to inherit from this class if it wants to be used as a response object - to do that, all an object has to do is implement `__view_result__`.
+    However, it's good convention to use this as the base class for objects that are solely used as a response.
+
+    It's probably not a good idea to reuse responses (i.e. return the same `Response` object from multiple calls), but it's not prevented.
+    """
 
     def __init__(
         self,
@@ -32,10 +46,24 @@ class Response(Generic[T]):
         headers: dict[str, str] | None = None,
         *,
         body_translate: BodyTranslateStrategy | None = _Find,
+        content_type: str = "text/plain"
     ) -> None:
+        """
+        Args:
+            body: Response body. This can be any type, and the `body_translate` parameter determines how it's converted into a string that view.py can read.
+            status: Status code to send to the user.
+            headers: Dictionary containing the response headers.
+            body_translate: This determines how the underlying `__view_result__` translates the body object into a string. By default, if the body object has a `__view_result__` attribute, this is set to `"result"`, and `"str"` otherwise. If this is `"custom"`, the `translate_body` method is called which can define any logic to translate the object. Note that if you're directly instantiating `Response`, you probably don't want to set this parameter - it's more for use from subclasses.
+            content_type: MIME type of the body. `"text/plain"` by default.
+        """
         self.body = body
+        """Implementation-specific body object. This will always be what's passed to the constructor of `Response`"""
         self.status = status
+        """Status code to be sent alongside the response."""
         self.headers = headers or {}
+        """Headers to be sent along with the response."""
+        self.content_type: str = content_type
+        """MIME type of the body. This is added to the headers last-minute in the `__view_result__` call, so you can update this value after it has been set."""
         self._raw_headers: list[tuple[bytes, bytes]] = []
 
         if body_translate:
@@ -43,11 +71,21 @@ class Response(Generic[T]):
         else:
             self.translate = "str" if not hasattr(body, "__view_result__") else "result"
 
-    def _custom(self, body: T) -> str:
+    def translate_body(self, body: T) -> str | bytes:
+        """
+        Translate the body via the `"custom"` body translate strategy. On classes that don't implement this, a `NotImplementedError` is raised.
+
+        Args:
+            body: Body object to translate to a string. This is dependent on the class.
+
+        Returns:
+            A `str` or `bytes` containing the translated body.
+        """
         raise NotImplementedError(
             'the "custom" translate strategy can only be used in subclasses that implement it'
         )  # noqa
 
+    @final
     def cookie(
         self,
         key: str,
@@ -62,7 +100,8 @@ class Response(Generic[T]):
         partitioned: bool = False,
         secure: bool = False,
     ) -> None:
-        """Set a cookie.
+        """
+        Set a cookie.
 
         Args:
             key: Cookie name.
@@ -73,7 +112,8 @@ class Response(Generic[T]):
             http_only: Whether the cookie should be HTTP only.
             same_site: SameSite setting for the cookie.
             partitioned: Whether to tie it to the top level site.
-            secure: Whether the cookie should enforce HTTPS."""
+            secure: Whether the cookie should enforce HTTPS.
+        """
         cookie_str = f"{key}={value}; SameSite={same_site}".encode()
 
         if expires:
@@ -106,15 +146,17 @@ class Response(Generic[T]):
         self._raw_headers.append((b"Set-Cookie", cookie_str))
 
     def _build_headers(self) -> tuple[tuple[bytes, bytes], ...]:
-        headers: list[tuple[bytes, bytes]] = [*self._raw_headers]
+        headers: list[tuple[bytes, bytes]] = [*self._raw_headers, (b"content-type", self.content_type.encode())]
 
         for k, v in self.headers.items():
             headers.append((k.encode(), v.encode()))
 
         return tuple(headers)
 
+    @final
     def __view_result__(self) -> ViewResult:
-        body: str = ""
+        """view.py response function. This should not be called manually, and it's implementation should be considered unstable, but this will always be here."""
+        body: str | bytes = ""
         if self.translate == "str":
             if isinstance(self.body, bytes):
                 body = self.body.decode()
@@ -123,7 +165,7 @@ class Response(Generic[T]):
         elif self.translate == "repr":
             body = repr(self.body)
         elif self.translate == "custom":
-            body = self._custom(self.body)
+            body = self.translate_body(self.body)
         else:
             view_result = getattr(self.body, "__view_result__", None)
 
@@ -142,7 +184,9 @@ class Response(Generic[T]):
 
 
 class HTML(Response[HTMLContent]):
-    """HTML response wrapper."""
+    """
+    HTML response wrapper.
+    """
 
     def __init__(
         self,
@@ -150,10 +194,9 @@ class HTML(Response[HTMLContent]):
         status: int = 200,
         headers: dict[str, str] | None = None,
     ) -> None:
-        super().__init__(body, status, headers, body_translate="custom")
-        self._raw_headers.append((b"content-type", b"text/html"))
+        super().__init__(body, status, headers, body_translate="custom", content_type="text/html")
 
-    def _custom(self, body: HTMLContent) -> str:
+    def translate_body(self, body: HTMLContent) -> str:
         parsed_body = ""
 
         if isinstance(body, Path):
@@ -174,12 +217,32 @@ class HTML(Response[HTMLContent]):
 
     @classmethod
     async def from_file(cls, path: str | Path) -> HTML:
+        """
+        Read an HTML file and load it as a response.
+        Roughly speaking, this shouldn't be used over `template()`, but it's a better option if you have a pure HTML file that doesn't need to get put through an extra template step.
+
+        Args:
+            path: `str` or `Path` object containing the path to the HTML file.
+
+        Example:
+            ```py
+            from view import HTML, get
+
+            @get("/")
+            async def index():
+                return await HTML.from_file("content/index.html")
+            ```
+        """
         async with aiofiles.open(path) as f:
             return cls(await f.read())
 
 
 class JSON(Response[Dict[str, Any]]):
-    """JSON response wrapper."""
+    """
+    JSON response wrapper.
+
+    Dictionaries passed to this object are serialized by `ujson.dumps`.
+    """
 
     def __init__(
         self,
@@ -187,10 +250,9 @@ class JSON(Response[Dict[str, Any]]):
         status: int = 200,
         headers: dict[str, str] | None = None,
     ) -> None:
-        super().__init__(body, status, headers, body_translate="custom")
-        self._raw_headers.append((b"content-type", b"application/json"))
+        super().__init__(body, status, headers, body_translate="custom", content_type="application/json",)
 
-    def _custom(self, body: dict[str, Any]) -> str:
+    def translate_body(self, body: dict[str, Any]) -> str:
         return ujson.dumps(body)