+++ src/view/templates.py
@@ -1,3 +1,13 @@
+"""
+view.py templating APIs
+
+This module contains all APIs related to rendering HTML through a template engine. As of this docstring being written, view.py supports the following engines:
+    - view.py's own template engine
+    - Jinja2
+    - Django's template engine
+    - Mako
+    - Chameleon
+"""
 from __future__ import annotations
 
 import inspect
@@ -39,7 +49,7 @@ sys.modules["_view_django"] = _DJANGO_HOOK
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "_view_django")
 
 
-class ViewRenderer:
+class _ViewRenderer:
     def __init__(self, parameters: dict[str, Any]):
         self.parameters = parameters
         self._last_if: bool | None = None
@@ -194,7 +204,30 @@ async def render(
     *,
     app: App | None = None,
 ) -> str:
-    """Render a template from the source instead of a filename. Generally should be used internally."""
+    """
+    Render a template from the source instead of a filename.
+    Generally should be used internally, but is considered stable.
+
+    This function does not require that an app has been created, but will attempt to get an app with `get_app()` regardless.
+    If `get_app()` fails, template engine instances are not stored.
+    
+    Args:
+        source: Source code to pass to the template engine.
+        engine: Template engine to use. Unlike `template()`, this does not try and load the default template engine from the config.
+        parameters: Variables to pass to the template engine for use in templates. By default, this will attempt to steal locals from the callers frame.
+        app: App to store engines on. If `None`, will attempt to call `get_app()`.
+
+    Example:
+        ```py
+        import aiofiles
+        from view import render
+
+
+        async def main():
+            async with aiofiles.open("template.html") as f:
+                html = await render(await f.read(), engine="jinja")
+        ```
+    """
     from .app import get_app
 
     if parameters is _CurrentFrame:
@@ -216,7 +249,7 @@ async def render(
         templaters = {}
 
     if engine == "view":
-        view = ViewRenderer(parameters or {})  # type: ignore
+        view = _ViewRenderer(parameters or {})  # type: ignore
         return await view.render(source)
     elif engine == "jinja":
         try:
@@ -279,7 +312,26 @@ async def template(
     app: App | None = None,
     **parameters: Any,
 ) -> HTML:
-    """Render a template with the specified engine. This returns a view.py HTML response."""
+    """
+    Render a template with the specified engine.
+
+    Args:
+        name: A string, which is appended the `.html` suffix (if it doesn't have it already), or a `Path` object, representing a file containing an HTML template.
+        directory: Directory to search for templates. If this is `None`, the directory specified in the configuration is used.
+        engine: Template engine to choose. If this is `None`, the default engine specified in the configuration is used.
+        frame: Frame to steal locals from for parameters. This will use the callers frame by default. If this is `None`, then no locals are sent to the template engine.
+        app: App instance to store template engine instances on. If this is `None`, `get_app` is called.
+        parameters: Extra values to send to the template engine as variables.
+
+    Example:
+        ```py
+        from view import get, template
+
+        @get("/")
+        async def index():
+            return await template("test")
+        ```
+    """
     from .app import get_app
 
     try:
@@ -328,7 +380,17 @@ async def markdown(
     directory: str | Path | None = _ConfigSpecified,
     app: App | None = None,
 ) -> HTML:
-    """Convert a markdown file into HTML. This returns a view.py HTML response."""
+    """
+    Convert a markdown file into HTML.
+
+    Args:
+        name: Equivalent to `name` in `template()`, with the exception of using the `.md` suffix instead of `.html`.
+        directory: Equivalent to `directory` in `template()`.
+        app: Equivalent to `app` in `template()`.
+
+    Raises:
+        NeedsDependencyError: `markdown` module is not installed.
+    """
     from .app import get_app
 
     try: